Qsim

Documentation for Qsim.

Qsim.basis

Qsim.ctrl_op

Qsim.qbit_op

Qsim.qbit_op

Qsim.basis — Method

function basis(qbits::BitVector)

Function to return a z-basis vector given a bitvector of qubit z-values .

Qsim.ctrl_op — Method

function ctrl_op(gate, cqbit::Integer, tqbit::Integer, qmax::Integer)

Returns a controlled version of gate with control and target bits defined by cqbit and tqbit in space with dimension 2^qmax.

Qsim.qbit_op — Method

qbit_op(gate, qbit::Integer, qmax::Integer)

Returns a qmax qubit operator with the provided gate acting on a single qubit.

Qsim.qbit_op — Method

function qubit_op(gate, qbits::Vector{Integer}, qmax)

Returns a qmax qubit operator with the provided gate acting on multiple qubits.